ai fixed bugs

+++ Server/Waves.c
@@ -27,14 +27,14 @@ uint32_t get_spawn_rarity(float difficulty)
 {
     if (difficulty < 1)
         difficulty = 1;
-    double rarity_seed = rr_frand();
+    long double rarity_seed = rr_frand();
     uint32_t rarity_cap = rr_rarity_id_common + (difficulty + 7) / 8;
     if (rarity_cap > rr_rarity_id_astral)
         rarity_cap = rr_rarity_id_astral;
     uint32_t rarity = rarity_cap >= rr_rarity_id_eternal ? rarity_cap - 3 : rarity_cap >= 2 ? rarity_cap - 2 : 0;
     for (; rarity < rarity_cap; ++rarity)
-        if (pow(1 - (1 - RR_MOB_WAVE_RARITY_COEFFICIENTS[rarity + 1]) * 0.3,
-                pow(1.5, difficulty)) >= rarity_seed)
+        if (powl(1 - (1 - RR_MOB_WAVE_RARITY_COEFFICIENTS[rarity + 1]) * 0.3L,
+                 powl(1.5L, (long double)difficulty)) >= rarity_seed)
             break;
     return rarity;
 }
